Authorization for Task

+++ app/Http/Controllers/TaskController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Task;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -77,9 +78,14 @@ class TaskController extends Controller
     public function edit($id)
     {
         $pageTitle = 'Edit Task';
-        $task = Task::find($id);
+        $task = Task::findOrFail($id);
 
-        return view('tasks.edit', ['pageTitle' => $pageTitle, 'task' => $task]);
+        Gate::authorize('update', $task);
+
+        return view('tasks.edit', [
+            'pageTitle' => $pageTitle,
+            'task' => $taskaaaaaaaa,
+        ]);
     }
 
     public function update(Request $request, $id)
@@ -93,7 +99,9 @@ class TaskController extends Controller
             $request->all()
         );
 
-        $task = Task::find($id);
+        $task = Task::findOrFail($id);
+
+        Gate::authorize('update', $task);
 
         $task->update([
             'name' => $request->name,
@@ -108,14 +116,18 @@ class TaskController extends Controller
     public function delete($id)
     {
         $title = 'Delete Task';
-        $task = Task::find($id);
+        $task = Task::findOrFail($id);
+
+        Gate::authorize('delete', $task);
 
         return view('tasks.delete', ['pageTitle' => $title, 'task' => $task]);
     }
 
     public function destroy($id)
     {
-        $task = Task::find($id);
+        $task = Task::findOrFail($id);
+        Gate::authorize('delete', $task);
+
         $task->delete();
         return redirect()->route('tasks.index');
     }
@@ -157,6 +169,8 @@ class TaskController extends Controller
     {
         $task = Task::findOrFail($id);
 
+        Gate::authorize('move', $task);
+
         $task->update([
             'status' => $request->status,
         ]);
@@ -167,6 +181,8 @@ class TaskController extends Controller
     public function complete(int $id)
     {
         $task = Task::findOrFail($id);
+
+        Gate::authorize('complete', $task);
 
         $task->update([
             'status' => Task::STATUS_COMPLETED,
